- JdbcOperations заменен на NamedParameterJdbcOperations. - Изменены запросы книг. - Добавление внешних ключей для таблицы книг. - Изменение сущности Book. - Изменение типа id.

+++ src/main/java/ru/otus/mkulikov/app/dao/AuthorDao.java
@@ -13,13 +13,13 @@ import java.util.List;
 
 public interface AuthorDao<T extends Author> {
 
-    T getById(int id);
+    T getById(long id);
 
     List<T> getAllObjects();
 
     int addObject(T t);
 
-    int deleteObject(int id);
+    int deleteObject(long id);
 
     int updateObject(T t);
 }
